refactor: migrate auth imports to domain structure and update related components
