+++ src/components/ui/Input.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { InputHTMLAttributes } from "react";
 import { cva, type VariantProps } from "class-variance-authority";
-import { mergeClasses } from "@/lib/tailwindUtils";
+import { mergeClasses } from "@/lib/utils/tailwind";
 import { Button, type ButtonProps } from "@/components/ui";
 import { Search } from "@/components/Icons";
