+++ src/components/ui/Tag.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { cva, type VariantProps } from "class-variance-authority";
-import { mergeClasses } from "@/lib/tailwindUtils";
+import { mergeClasses } from "@/lib/utils/tailwind";
 
 const tagVariants = cva("inline-flex items-center rounded-full font-medium w-fit gap-2", {
 	variants: {
